WIP for adding style guide
This time:

* Integrate `react-styleguidist`
* Move Redux components out of `src/components` so styleguidist doesn't complain.
* Add styleguidist config
* Add `Alert` and `AlertContainer` components

Next up:

* Add documentation
