+++ src/App.js
@@ -10,8 +10,8 @@ import {
 import { PersistGate } from 'redux-persist/lib/integration/react';
 import { Provider } from 'react-redux';
 
-import ConnectedFontWeightTester from 'components/ConnectedFontWeightTester';
-import ConnectedTypefaceTester from 'components/ConnectedTypefaceTester';
+import ConnectedFontWeightTester from 'state-aware-components/ConnectedFontWeightTester';
+import ConnectedTypefaceTester from 'state-aware-components/ConnectedTypefaceTester';
 import Home from 'components/Home';
 import logo from 'images/logo.svg';
 import React, { Component } from 'react';
