+++ src/state-aware-components/ConnectedFontWeightTester.test.js
@@ -1,8 +1,4 @@
-import {
-  addGlobalSetting,
-  addSettingForPage,
-} from 'actions/settings';
-
+import { addGlobalSetting } from 'actions/settings';
 import {
   cleanUpTests,
   initTests,
@@ -11,9 +7,9 @@ import {
 import {
   mapDispatchToProps,
   mapStateToProps,
-} from 'components/ConnectedTypefaceTester';
+} from 'state-aware-components/ConnectedFontWeightTester';
 
-describe('ConnectedTypefaceTester', () => {
+describe('ConnectedFontWeightTester', () => {
   let sandbox;
   let dispatchStub;
 
@@ -37,25 +33,10 @@ describe('ConnectedTypefaceTester', () => {
       const props = mapStateToProps({
         settings: {
           family: 'Averta Std',
-          typefaceTester: {
-            snippet: 'tale-of-two-cities',
-          },
         },
       });
 
       expect(props.family).to.equal('Averta Std');
-      expect(props.snippet).to.equal('tale-of-two-cities');
-    });
-
-    it('should handle when there are no settings for the page', () => {
-      const props = mapStateToProps({
-        settings: {
-          family: 'Averta Std',
-        },
-      });
-
-      expect(props.family).to.equal('Averta Std');
-      expect(props.snippet).to.be.undefined();
     });
   });
 
@@ -69,8 +50,6 @@ describe('ConnectedTypefaceTester', () => {
     it('should return an object of the right form', () => {
       expect(props.onUpdateFamily).to.exist();
       expect(props.onUpdateFamily).to.be.a('function');
-      expect(props.onUpdateText).to.exist();
-      expect(props.onUpdateText).to.be.a('function');
     });
 
     it('should call dispatch when onUpdateFamily is invoked', () => {
@@ -78,15 +57,6 @@ describe('ConnectedTypefaceTester', () => {
 
       expect(dispatchStub.callCount).to.equal(1);
       expect(dispatchStub.calledWith(addGlobalSetting('family', 'Brix Slab'))).to.be.true();
-    });
-
-    it('should call dispatch when onUpdateText is invoked', () => {
-      props.onUpdateText('scandal-in-bohemia');
-
-      expect(dispatchStub.callCount).to.equal(1);
-      expect(dispatchStub.calledWith(
-        addSettingForPage('typefaceTester', 'snippet', 'scandal-in-bohemia')
-      )).to.be.true();
     });
   });
 });
